refactor: move to LaunchButton.tsx
